Implement parallel write-safety
Technically the extension is *already* parallel-read-safe because it's
not implicated in reading at all, its "reading" work occurs during
writing.

As preparatory code clean up the code a bit:

* remove `app.locales` entirely, the locales fetching only occurs
  during the final writing so `get_locales` can return the information
* update some etree calls to use kwargs

For the multiprocessing, `add_html_link` can be called from different
worker processes while `create_sitemap` is always called from a single
one. While the env is fanned out when creating workers (so they can
all see the env) unlike reading there is no merge step for
writing. Therefore an other synchronisation mechanism is needed to
feed worker information back to the parent.

`multiprocessing.Queue` didn't work (it complains about inheritance
between processes), while likely quite a bit heavier
`multiprocessing.Manager.Queue` does, so use that: set the queue on
the `env`, fill the queue on `add_html_link`, and pull from the queue
until it's empty when generating the sitemap.

One unfortunate issue is the deoptimisation of serial writing: a
managed queue is always created which is much heavier than a list (in
my understanding the manager is essentially a separate process which
manages synchronisation between the various users of the
proxy). Conditionally using a managed queue seems difficult: there is
a `Builder.parallel_ok` flag but it's only set during `build()`, so
not set at creation. We could hand-roll the corresponding checks, but
they're mostly implementation details which is not convenient.